app/requirements.txt
Flask==2.3.4
kafka-python==2.0.2
psycopg2-binary==2.9.9
opencv-python==4.8.1.78
scikit-learn==1.3.0
joblib==1.3.2
pandas==2.0.3
